make cloud.Metadata avaialable to external clients
