+++ cloud/cloud.go
@@ -1,0 +1,10 @@
+// Package cloud contains app-related functionality.
+package cloud
+
+// Metadata contains app-related information about the robot.
+type Metadata struct {
+	PrimaryOrgID  string
+	LocationID    string
+	MachineID     string
+	MachinePartID string
+}
