+++ internal/cloud/service.go
@@ -114,11 +114,3 @@ func (cm *cloudManagedService) Close(ctx context.Context) error {
 // ErrNotCloudManaged is returned if a connection is requested but the robot is not
 // yet cloud managed.
 var ErrNotCloudManaged = errors.New("this robot is not cloud managed")
-
-// Metadata contains app-related information about the robot.
-type Metadata struct {
-	PrimaryOrgID  string
-	LocationID    string
-	MachineID     string
-	MachinePartID string
-}
